tests: move __init__ to common
Treat tests as a directory full of modules, rather than as a package.
This will allow us to stop adding the root of the repo to the Python path.
